requirements_testing.txt
-r requirements.txt
pytest==6.1
pytest-django==4.1.0
factory_boy==3.2.0
Faker==8.10.0